streamlit/requirements.txt
streamlit==1.27.0
pandas==2.0.3
plotly==5.16.1
cassandra-driver==3.28.0
matplotlib==3.7.2